requirements.txt
coinbase==2.0.0
defusedxml==0.4.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.3
django-allauth==0.23.0
django-debug-toolbar==1.3.2
django-endless-pagination==2.0
django-facebook==6.0.3
django-toolbelt==0.0.1
gunicorn==19.3.0
oauthlib==1.0.1
Pillow==2.9.0
psycopg2==2.6.1
python3-openid==3.0.6
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
sqlparse==0.1.16
static3==0.6.1
Unidecode==0.4.18
wheel==0.24.0
